Fix PS5.1-incompatible use of IsPathFullyQualified
The Initialize-PSBuild function was not compatible with .NET Framework / PowerShell 5.1 due to the use of the System.IO.Path.IsPathFullyQualified method. I changed the function to check if the value of OutDir already started with BHProjectPath rather than testing if OutDir was a fully qualified path on it's own. This seems to have solved the issue and tests are passing.